backend/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
scikit-fuzzy==0.4.2
scipy==1.14.0
six==1.16.0
Werkzeug==3.0.3
